+++ index.html
@@ -4,7 +4,10 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>CS Department | University Of Ghana</title>
-    <script src="https://kit.fontawesome.com/e9f5458810.js" crossorigin="anonymous"></script>
+    <script
+      src="https://kit.fontawesome.com/e9f5458810.js"
+      crossorigin="anonymous"
+    ></script>
     <link rel="preconnect" href="https://fonts.googleapis.com" />
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
     <link
@@ -13,7 +16,6 @@
     />
     <link rel="stylesheet" href="./assets/css/index.css" />
     <link rel="stylesheet" href="./assets/css/mediaquery.css" />
-
   </head>
   <body>
     <!-- Navigation bar starts here -->
@@ -36,8 +38,12 @@
             <a href="./templates/news.html"> News </a>
           </li>
           <li>
-            <a href="./index.html"> 
-                <img src="./assets/images/ug_logo.png" alt="ug_log" id="_logo_ug">
+            <a href="./index.html">
+              <img
+                src="./assets/images/ug_logo.png"
+                alt="ug_log"
+                id="_logo_ug"
+              />
             </a>
           </li>
           <li>
@@ -54,20 +60,48 @@
           </li>
         </ul>
       </nav>
-      <!-- desktop
-      <nav></nav> -->
-      <button type="button" id="search-btn">
+      <a href="./index.html">
+        <img src="./assets/images/ug_logo.png" alt="ug_log" id="mb_ug_logo" />
+      </a>
+      <nav id="mobile-nav" class="nav">
+        <ul>
+          <button id="close-btn" type="button">
+            <i class="fa-solid fa-xmark"></i>
+          </button>
+          <li>
+            <a href="./index.html"> Home </a>
+          </li>
+          <li>
+            <a href="./templates/about.html"> About </a>
+          </li>
+          <li>
+            <a href="./templates/contacts.html"> Contact </a>
+          </li>
+          <li>
+            <a href="./templates/news.html"> News </a>
+          </li>
 
+          <li>
+            <a href="./templates/events.html"> events </a>
+          </li>
+          <li>
+            <a href="./templates/researchs.html"> Research </a>
+          </li>
+          <li>
+            <a href="./templates/catalog.html"> Catalogs </a>
+          </li>
+          <li>
+            <a href="./templates/resources.html"> Resources </a>
+          </li>
+        </ul>
+      </nav>
+      <button type="button" id="search-btn">
         <i class="fa-solid fa-magnifying-glass"></i>
       </button>
     </header>
-    <section id="hero">
-
-    </section>
-    <section id="hero">
-      
-    </section>
+    <section id="hero"></section>
+    <section id="hero"></section>
     <!-- Navigation bar ends here -->
-    <script src="./index.js" type="module"></script><script src="https://kit.fontawesome.com/e9f5458810.js" crossorigin="anonymous"></script>
+    <script src="./index.js" type="module"></script>
   </body>
 </html>
